+++ src/internal.c
@@ -9,6 +9,7 @@
 #include "internal/expr-interp.c"
 #include "internal/expr-interp-rotate.c"
 #include "internal/fn.c"
+#include "internal/hash.c"
 #include "internal/internal.c"
 #include "internal/nse-defuse.c"
 #include "internal/quo.c"